--- a/v2/binding/binary_writer.go
+++ b/v2/binding/binary_writer.go
@@ -3,8 +3,6 @@ package binding
 import (
 	"context"
 	"io"
-
-	"github.com/cloudevents/sdk-go/v2/binding/spec"
 )
 
 // BinaryWriter is used to visit a binary Message and generate a new representation.
@@ -14,20 +12,10 @@ import (
 //
 // Start() and End() methods are invoked every time this BinaryWriter implementation is used to visit a Message
 type BinaryWriter interface {
+	MessageMetadataWriter
+
 	// Method invoked at the beginning of the visit. Useful to perform initial memory allocations
 	Start(ctx context.Context) error
-
-	// Set a standard attribute.
-	//
-	// The value can either be the correct golang type for the attribute, or a canonical
-	// string encoding. See package types to perform the needed conversions
-	SetAttribute(attribute spec.Attribute, value interface{}) error
-
-	// Set an extension attribute.
-	//
-	// The value can either be the correct golang type for the attribute, or a canonical
-	// string encoding. See package types to perform the needed conversions
-	SetExtension(name string, value interface{}) error
 
 	// SetData receives an io.Reader for the data attribute.
 	// io.Reader is not invoked when the data attribute is empty