--- a/v2/binding/buffering/binary_buffer_message.go
+++ b/v2/binding/buffering/binary_buffer_message.go
@@ -16,6 +16,7 @@ var binaryMessagePool bytebufferpool.Pool
 // binaryBufferedMessage implements a binary-mode message as a simple struct.
 // This message implementation is used by CopyMessage and BufferMessage
 type binaryBufferedMessage struct {
+	version    spec.Version
 	metadata   map[spec.Attribute]interface{}
 	extensions map[string]interface{}
 	body       *bytebufferpool.ByteBuffer
@@ -85,6 +86,7 @@ func (m *binaryBufferedMessage) SetData(data io.Reader) error {
 
 func (m *binaryBufferedMessage) SetAttribute(attribute spec.Attribute, value interface{}) error {
 	// If spec version we need to change to right context struct
+	m.version = attribute.Version()
 	m.metadata[attribute] = value
 	return nil
 }
@@ -98,5 +100,18 @@ func (m *binaryBufferedMessage) End(ctx context.Context) error {
 	return nil
 }
 
+func (m *binaryBufferedMessage) GetAttribute(k spec.Kind) (spec.Attribute, interface{}) {
+	a := m.version.AttributeFromKind(k)
+	if a != nil {
+		return a, m.metadata[a]
+	}
+	return nil, nil
+}
+
+func (m *binaryBufferedMessage) GetExtension(name string) interface{} {
+	return m.extensions[name]
+}
+
 var _ binding.Message = (*binaryBufferedMessage)(nil) // Test it conforms to the interface
+var _ binding.MessageMetadataReader = (*binaryBufferedMessage)(nil)
 var _ binding.BinaryWriter = (*binaryBufferedMessage)(nil)