--- a/v2/binding/event_message.go
+++ b/v2/binding/event_message.go
@@ -60,6 +60,20 @@ func (m *EventMessage) ReadBinary(ctx context.Context, b BinaryWriter) (err erro
 	return b.End(ctx)
 }
 
+func (m *EventMessage) GetAttribute(k spec.Kind) (spec.Attribute, interface{}) {
+	sv := spec.VS.Version(m.Context.GetSpecVersion())
+	a := sv.AttributeFromKind(k)
+	if a != nil {
+		return a, a.Get(m.Context)
+	}
+	return nil, nil
+}
+
+func (m *EventMessage) GetExtension(name string) interface{} {
+	ext, _ := m.Context.GetExtension(name)
+	return ext
+}
+
 func eventContextToBinaryWriter(c event.EventContext, b BinaryWriter) (err error) {
 	// Pass all attributes
 	sv := spec.VS.Version(c.GetSpecVersion())
@@ -84,7 +98,8 @@ func eventContextToBinaryWriter(c event.EventContext, b BinaryWriter) (err error
 
 func (*EventMessage) Finish(error) error { return nil }
 
-var _ Message = (*EventMessage)(nil) // Test it conforms to the interface
+var _ Message = (*EventMessage)(nil)               // Test it conforms to the interface
+var _ MessageMetadataReader = (*EventMessage)(nil) // Test it conforms to the interface
 
 // UseFormatForEvent configures which format to use when marshalling the event to structured mode
 func UseFormatForEvent(ctx context.Context, f format.Format) context.Context {