--- a/v2/binding/to_event.go
+++ b/v2/binding/to_event.go
@@ -20,7 +20,7 @@ var ErrCannotConvertToEvent = errors.New("cannot convert message to event")
 // This function returns the Event generated from the Message and the original encoding of the message or
 // an error that points the conversion error.
 // transformers can be nil and this function guarantees that they are invoked only once during the encoding process.
-func ToEvent(ctx context.Context, message MessageReader, transformers ...TransformerFactory) (*event.Event, error) {
+func ToEvent(ctx context.Context, message MessageReader, transformers ...Transformer) (*event.Event, error) {
 	if message == nil {
 		return nil, nil
 	}
@@ -32,11 +32,7 @@ func ToEvent(ctx context.Context, message MessageReader, transformers ...Transfo
 			switch mt := m.(type) {
 			case *EventMessage:
 				e := (*event.Event)(mt)
-				var tf TransformerFactories = transformers
-				if err := tf.EventTransformer()(e); err != nil {
-					return nil, err
-				}
-				return e, nil
+				return e, Transformers(transformers).Transform(mt, (*messageToEventBuilder)(e))
 			case MessageWrapper:
 				m = mt.GetWrappedMessage()
 			default:
@@ -47,36 +43,31 @@ func ToEvent(ctx context.Context, message MessageReader, transformers ...Transfo
 	}
 
 	e := event.New()
-	encoder := &messageToEventBuilder{event: &e}
-	if _, err := DirectWrite(
+	encoder := (*messageToEventBuilder)(&e)
+	_, err := DirectWrite(
 		context.Background(),
 		message,
 		encoder,
 		encoder,
-	); err != nil {
+	)
+	if err != nil {
 		return nil, err
 	}
-	var tf TransformerFactories = transformers
-	if err := tf.EventTransformer()(&e); err != nil {
-		return nil, err
-	}
-	return &e, nil
+	return &e, Transformers(transformers).Transform((*EventMessage)(&e), encoder)
 }
 
-type messageToEventBuilder struct {
-	event *event.Event
-}
+type messageToEventBuilder event.Event
 
 var _ StructuredWriter = (*messageToEventBuilder)(nil)
 var _ BinaryWriter = (*messageToEventBuilder)(nil)
 
-func (b *messageToEventBuilder) SetStructuredEvent(ctx context.Context, format format.Format, event io.Reader) error {
+func (b *messageToEventBuilder) SetStructuredEvent(ctx context.Context, format format.Format, ev io.Reader) error {
 	var buf bytes.Buffer
-	_, err := io.Copy(&buf, event)
+	_, err := io.Copy(&buf, ev)
 	if err != nil {
 		return err
 	}
-	return format.Unmarshal(buf.Bytes(), b.event)
+	return format.Unmarshal(buf.Bytes(), (*event.Event)(b))
 }
 
 func (b *messageToEventBuilder) Start(ctx context.Context) error {
@@ -94,7 +85,7 @@ func (b *messageToEventBuilder) SetData(data io.Reader) error {
 		return err
 	}
 	if w != 0 {
-		b.event.DataEncoded = buf.Bytes()
+		b.DataEncoded = buf.Bytes()
 	}
 	return nil
 }
@@ -108,15 +99,15 @@ func (b *messageToEventBuilder) SetAttribute(attribute spec.Attribute, value int
 		}
 		switch str {
 		case event.CloudEventsVersionV03:
-			b.event.Context = b.event.Context.AsV03()
+			b.Context = b.Context.AsV03()
 		case event.CloudEventsVersionV1:
-			b.event.Context = b.event.Context.AsV1()
+			b.Context = b.Context.AsV1()
 		default:
 			return fmt.Errorf("unrecognized event version %s", str)
 		}
 		return nil
 	}
-	return attribute.Set(b.event.Context, value)
+	return attribute.Set(b.Context, value)
 }
 
 func (b *messageToEventBuilder) SetExtension(name string, value interface{}) error {
@@ -124,6 +115,5 @@ func (b *messageToEventBuilder) SetExtension(name string, value interface{}) err
 	if err != nil {
 		return err
 	}
-	b.event.SetExtension(name, value)
-	return nil
+	return b.Context.SetExtension(name, value)
 }