--- a/v2/protocol/amqp/options.go
+++ b/v2/protocol/amqp/options.go
@@ -49,7 +49,7 @@ func WithReceiverLinkOption(opt amqp.LinkOption) Option {
 type SenderOptionFunc func(sender *sender)
 
 // WithTransformer adds a transformer, which Sender uses while encoding a binding.Message to an amqp.Message
-func WithTransformer(transformer binding.TransformerFactory) SenderOptionFunc {
+func WithTransformer(transformer binding.Transformer) SenderOptionFunc {
 	return func(sender *sender) {
 		sender.transformers = append(sender.transformers, transformer)
 	}