… v2/binding/buffering/copy_message_benchmark_test.go
… v2/binding/buffering/copy_message_test.go
… v2/binding/event_message_test.go
… v2/binding/test/mock_binary_message.go
… v2/binding/test/mock_transformer.go
… v2/binding/test/test.go
… v2/binding/test/transformer.go
… v2/binding/to_event_test.go
… v2/binding/transformer/add_metadata_test.go
… v2/binding/transformer/default_metadata.go
… v2/binding/transformer/default_metadata_test.go
… v2/binding/transformer/delete_metadata.go
… v2/binding/transformer/delete_metadata_test.go
… v2/binding/transformer/set_metadata.go
… v2/binding/transformer/set_metadata_test.go
… v2/binding/transformer/update_metadata.go
… v2/binding/transformer/update_metadata_test.go
… v2/binding/transformer/version.go
… v2/binding/transformer/version_test.go
… v2/binding/write.go
… v2/protocol/amqp/message.go
… v2/protocol/amqp/sender.go
… v2/protocol/amqp/write_message.go
… v2/protocol/http/headers.go
… v2/protocol/http/message.go
… v2/protocol/http/options_test.go
… v2/protocol/http/protocol.go
… v2/protocol/http/protocol_test.go
… v2/protocol/http/write_request.go
… v2/protocol/http/write_responsewriter.go